requirements.txt
openai
python-dotenv
pytest
fastapi
uvicorn[standard]
python-multipart
